Merging develop into master for release May 8th (#8)
* Added a reference

* Update README.md

* Added Basic Java Project structure

* Updated fileUtil object

The fileUtil object will allow for a file to be defined when you instantiate the object.

* Jenkins test commit

* Jenkins build test 2

* Added build badge into ReadeMe doc

* Added Project Requirement section to the ReadeMe.

Added Project Requirement section to the ReadeMe doc.

* Added a new class to represent a Node in the QuadTree.

The QuadNode contains properties for the x and y coordinates, the data, and four inner quad nodes (to represent each region of the graph).
The QuadNode contains behaviors for getting x/y coordinates, getting the data, getting the inner quad nodes, setting x/y coordinates, setting the data, and setting the inner quad nodes.

* Added insertion and bounds checking behaviors.

Added essential behaviors to the QuadTree class. Also added code to instantiate a new QuadTree in the tester and insert a couple of datapoints. Made a slight change in the QuadNode class relating to the names of variables to be more intent revealing.

* Removed the build status.

We shouldn't link to the login page of our build server. It should be private information.

* Removed unused files.

DynamoDB will act like the datastructure, so we will not actually need a Node class. All we will need is for the QuadTree to generate output given a grid

* Updated the QuadTree to be able to generate output given a grid

Added a generateQuadTree method that will generate output based on the scale of the QuadTree

* Fixed some bugs and added enhancements

Fixed a StackOverFlow error that was the result of never hitting the base case and an algebraic error. Also added the ability for the user to define the gap (default=1).

* Updated the formatting and design.

Updated the formatting of some code - also provided more consistent naming, encapsulated some logic, and fixed an important naming error.

* Corrected a typo.

Changed the order of coordinates when printing the syntax/format of how the node information is displayed.

* Updated the instructions
Added information about running the tester with user-defined coordinates.

* Added a new behavior

Added the behavior to calculate the total amount of nodes that will be generated given a maximum coordinate like maxX and maxY.

* Updated the modifier.

There is no need for the getTotalNodes behavior to be public - I have changed it to private.

* Added behaviors to output the QuadTree to files.

Added behaviors to the FileUtil that allows for buildint a String and then writing it to a file. Instantiated and used that object inside of QuadTree when generating the QuadTree.

* Fixed an error

Corrected a bug that prevented the (numberOfFiles-1)th file from being generated.

* Removed unneeded line of code

Removed a System.out.println(...) that was used for debugging.

* Updated the format of output files

The format of the output files is dependent on the user defined file extension. If csv detected, then headers and tabs will be added as necessary.

* Update README.md

* v1 of the project cleaned up + Add docs to each projects

* Refactored tester file to apply simpler && unified execution

* QuadTree general constructors added + refacotred naming on v1 tree

* QuadTree validation check added

* Refacotred indentation + var name + stubbed generate method added

* Refactored variables into TreeNode object

* Generates tree in correct Quadrant order + refactoring code

* Logic refactored to display binary id + finds leaf nodes

* Added file output

Updated the QuadTree class to interact with the FileUtil class in order to build csv files. Some logic was added to only add in leaf nodes when building a file. Additionally, indentation was updated to be a 4 space indent. Imports are now package specific instead of using wildcards. Also added some setters for setting the partition-limit, filename-prefix, and filename-extension. Defaults to 10, "aws", and "csv" respectively.

* Set the stage for S3

Created the ant buildfile so that we can compile Java files that use the aws-java-sdk from terminal. Added the S3 packages to QuadTree.java and it will successfully compile from command. Just run ant && ant clean

* Updated gitignore

Update to ignore the vendor path. The vendor path will contain third party tools and SDKs, which will not be included on the repo.

* Removed vendor

Removed the vendor directory, which we will not be including on the repo.

* Added an install file

Added a shell script to help the user install the Amazon SDK that the java files require.

* Updated the install file

Changed the shebang to use /bin/sh, so that when the shell script is run directly, it will execute with sh.

* Updated the permissions

Removed execute permission - the user should manually add +x if they trust the install script and wish to execute it.

* Connected the QuadTree with S3

Implemented making a connection to S3: will create a bucket if one does not exist, will upload files to that bucket as they are generated, and will print all of the objects from the bucket once the QuadTree is generated.

* Updated the bucket-name

Changed the bucket-name from "ewu.quadtree.s3" to "quadtree-main-bucket"

* Updated the bucket name

Changed the S3 bucket name from "ewu.quadtree.s3" to "tf-quadtree-main-bucket"

* Updated the instructions

Added information about how to use apache ant to compile and run our project.